+++ example/lib/constants.dart
@@ -32,4 +32,12 @@ class Constants {
       "https://imgix.bustle.com/uploads/image/2020/8/5/23905b9c-6b8c-47fa-bc0f-434de1d7e9bf-avengers-5.jpg";
   static String elephantDreamStreamUrl =
       "http://cdn.theoplayer.com/video/elephants-dream/playlist.m3u8";
+  static String tokenEncodedHlsUrl =
+      "https://amssamples.streaming.mediaservices.windows.net/830584f8-f0c8-4e41-968b-6538b9380aa5/TearsOfSteelTeaser.ism/manifest(format=m3u8-aapl)";
+  static String tokenEncodedHlsToken =
+      "Bearer=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1cm46bWljcm9zb2Z0OmF6dXJlOm1lZGlhc2VydmljZXM6Y29udGVudGtleWlkZW50aWZpZXIiOiI5ZGRhMGJjYy01NmZiLTQxNDMtOWQzMi0zYWI5Y2M2ZWE4MGIiLCJpc3MiOiJodHRwOi8vdGVzdGFjcy5jb20vIiwiYXVkIjoidXJuOnRlc3QiLCJleHAiOjE3MTA4MDczODl9.lJXm5hmkp5ArRIAHqVJGefW2bcTzd91iZphoKDwa6w8";
+  static String widevineVideoUrl =
+      "https://storage.googleapis.com/wvmedia/cenc/h264/tears/tears_sd.mpd";
+  static String widevineLicenseUrl =
+      "https://proxy.uat.widevine.com/proxy?provider=widevine_test";
 }