+++ lib/src/configuration/better_player_drm_type.dart
@@ -1,0 +1,4 @@
+///Types of available DRM's
+///token -> supported for both iOS/Android
+///widevine -> supported only for Android
+enum BetterPlayerDrmType { token, widevine }